refactoring class
