+++ Example/AndesUI/ProgressIndicator/Views/ProgressIndicatorViewController.xib
@@ -41,6 +41,10 @@
                                         <userDefinedRuntimeAttribute type="color" keyPath="tint">
                                             <color key="value" systemColor="systemPinkColor" red="1" green="0.1764705882" blue="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                         </userDefinedRuntimeAttribute>
+                                        <userDefinedRuntimeAttribute type="string" keyPath="label" value="Large"/>
+                                        <userDefinedRuntimeAttribute type="color" keyPath="textColor">
+                                            <color key="value" systemColor="systemRedColor" red="1" green="0.23137254900000001" blue="0.18823529410000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                                        </userDefinedRuntimeAttribute>
                                     </userDefinedRuntimeAttributes>
                                 </view>
                                 <view contentMode="scaleToFill" placeholderIntrinsicWidth="infinite" placeholderIntrinsicHeight="48" insetsLayoutMarginsFromSafeArea="NO" translatesAutoresizingMaskIntoConstraints="NO" id="VQW-ad-oTb" customClass="AndesProgressIndicator" customModule="AndesUI">
@@ -50,6 +54,9 @@
                                         <userDefinedRuntimeAttribute type="string" keyPath="ibSize" value="MEDIUM"/>
                                         <userDefinedRuntimeAttribute type="color" keyPath="tint">
                                             <color key="value" systemColor="systemPurpleColor" red="0.68627450980000004" green="0.32156862749999998" blue="0.87058823529999996" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                                        </userDefinedRuntimeAttribute>
+                                        <userDefinedRuntimeAttribute type="color" keyPath="textColor">
+                                            <color key="value" systemColor="systemOrangeColor" red="1" green="0.58431372550000005" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                         </userDefinedRuntimeAttribute>
                                     </userDefinedRuntimeAttributes>
                                 </view>
@@ -61,6 +68,7 @@
                                         <userDefinedRuntimeAttribute type="color" keyPath="tint">
                                             <color key="value" systemColor="systemPinkColor" red="1" green="0.1764705882" blue="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                         </userDefinedRuntimeAttribute>
+                                        <userDefinedRuntimeAttribute type="string" keyPath="label" value="Small"/>
                                     </userDefinedRuntimeAttributes>
                                 </view>
                             </subviews>
@@ -83,7 +91,7 @@
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="9mE-e9-I08" secondAttribute="trailing" id="t7f-5n-xx9"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
-            <point key="canvasLocation" x="131.8840579710145" y="84.375"/>
+            <point key="canvasLocation" x="-348" y="98"/>
         </view>
     </objects>
 </document>
