+++ LibraryComponents/Classes/Core/AndesProgressIndicator/Size/AndesProgressIndicatorSizeMedium.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 internal struct AndesProgressIndicatorSizeMedium: AndesProgressIndicatorSizeProtocol {
-
+    var textSpacing: CGFloat = 20
     var strokeWidth: CGFloat = 3
     var height: CGFloat = 32
