--- LibraryComponents/Classes/Core/AndesProgressIndicator/View/Default/AndesProgressIndicatorViewDefault.xib
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="16097.3" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16097.3"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="AndesProgressIndicatorViewDefault" customModule="AndesUI">
-            <connections>
-                <outlet property="componentView" destination="iN0-l3-epB" id="pJc-XA-8DK"/>
-                <outlet property="containerView" destination="Bhg-eP-9Nc" id="Df5-FI-ZTM"/>
-                <outlet property="containerViewHeightConstraint" destination="FKk-BF-3wc" id="iPq-fw-IL2"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="221" height="166"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Bhg-eP-9Nc" customClass="AndesCircularProgressBar" customModule="AndesUI">
-                    <rect key="frame" x="86.5" y="59" width="48" height="48"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="48" id="FKk-BF-3wc"/>
-                        <constraint firstAttribute="width" secondItem="Bhg-eP-9Nc" secondAttribute="height" multiplier="1:1" id="tNK-eR-NcI"/>
-                    </constraints>
-                    <userDefinedRuntimeAttributes>
-                        <userDefinedRuntimeAttribute type="number" keyPath="ringWidth">
-                            <real key="value" value="5"/>
-                        </userDefinedRuntimeAttribute>
-                        <userDefinedRuntimeAttribute type="color" keyPath="color">
-                            <color key="value" systemColor="systemTealColor" red="0.35294117650000001" green="0.7843137255" blue="0.98039215690000003" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                        </userDefinedRuntimeAttribute>
-                    </userDefinedRuntimeAttributes>
-                </view>
-            </subviews>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <constraints>
-                <constraint firstItem="Bhg-eP-9Nc" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="0wg-L1-ftp"/>
-                <constraint firstItem="Bhg-eP-9Nc" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="iN0-l3-epB" secondAttribute="leading" id="K4o-7Y-PXA"/>
-                <constraint firstItem="Bhg-eP-9Nc" firstAttribute="top" relation="greaterThanOrEqual" secondItem="iN0-l3-epB" secondAttribute="top" id="Un0-fG-fAg"/>
-                <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="Bhg-eP-9Nc" secondAttribute="bottom" id="eFF-DI-225"/>
-                <constraint firstItem="Bhg-eP-9Nc" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="oeu-aM-2Vd"/>
-                <constraint firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="Bhg-eP-9Nc" secondAttribute="trailing" id="u67-Lm-g5Z"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="-141.30434782608697" y="-13.392857142857142"/>
-        </view>
-    </objects>
-</document>
