+++ LibraryComponents/Classes/Core/AndesProgressIndicator/View/Default/Circular/AndesProgressIndicatorViewCircular.swift
@@ -1,0 +1,23 @@
+//
+//  
+//  AndesProgressIndicatorViewDefault.swift
+//  AndesUI
+//
+//  Created by Juan Andres Vasquez Ferrer on 30-11-20.
+//
+//
+
+import UIKit
+
+class AndesProgressIndicatorViewCircular: AndesProgressIndicatorAbstractView {
+
+    override func loadNib() {
+        let bundle = AndesBundle.bundle()
+        bundle.loadNibNamed("AndesProgressIndicatorViewCircular", owner: self, options: nil)
+    }
+
+    override func updateView() {
+        super.updateView()
+        self.containerView.progress = 1
+    }
+}
